ASTON-18 [refactor]: decomposition
