+++ src/components/characterContainer/characterContainer.module.scss
@@ -1,0 +1,9 @@
+.page {
+  &__list {
+    margin-top: 65px;
+  }
+
+  &__section {
+    width: 100%;
+  }
+}
