+++ src/components/characterContainer/index.ts
@@ -1,0 +1,1 @@
+export * from './characterContainer'
