+++ src/page/search-page/search-page.module.scss
@@ -1,8 +1,4 @@
 .page {
-  &__list {
-    margin-top: 65px;
-  }
-
   &__section {
     width: 100%;
